+++ src/main/java/moze_intel/projecte/gameObjs/items/HyperkineticLens.java
@@ -36,7 +36,7 @@ public class HyperkineticLens extends ItemCharge implements IProjectileShooter
 	public boolean shootProjectile(EntityPlayer player, ItemStack stack) 
 	{
 		World world = player.worldObj;
-		int requiredEmc = Constants.EXPLOSIVE_LENS_COST[this.getCharge(stack)];
+		long requiredEmc = Constants.EXPLOSIVE_LENS_COST[this.getCharge(stack)];
 		
 		if (!consumeFuel(player, stack, requiredEmc, true))
 		{
